Started pulling in the players location correctly, well sort of, also started working on rendering only vicinity onto the map on load.

+++ api/models/Population.js
@@ -14,7 +14,7 @@ module.exports = {
   username: {
       model: 'user',
     },
-  amount: 'float'
+  amount: 'int'
   }
 };
